Mirai for Python/requirements.txt
dashscope
yiri-mirai-rc
